10-10dx | Remove metadata select for adoption papers file field
